# backend/apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "香港诊所保险保单智能填报 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/doctor/insurance-companies/{company_id}/templates": {
      "get": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "List Templates",
        "operationId": "list_templates_api_doctor_insurance_companies__company_id__templates_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateBrief"
                  },
                  "title": "Response List Templates Api Doctor Insurance Companies  Company Id  Templates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doctor/claims": {
      "post": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "Create Claim",
        "operationId": "create_claim_api_doctor_claims_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "List Claims",
        "operationId": "list_claims_api_doctor_claims_get",
        "parameters": [
          {
            "name": "patient_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Name"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ClaimListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doctor/claims/{claim_id}/medical-record": {
      "put": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "Submit Medical Record",
        "operationId": "submit_medical_record_api_doctor_claims__claim_id__medical_record_put",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Claim Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecordSubmit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doctor/claims/{claim_id}/fields": {
      "put": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "Update Fields",
        "operationId": "update_fields_api_doctor_claims__claim_id__fields_put",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Claim Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doctor/claims/{claim_id}/confirm": {
      "post": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "Confirm",
        "operationId": "confirm_api_doctor_claims__claim_id__confirm_post",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Claim Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doctor/claims/{claim_id}/generate-pdf": {
      "post": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "Generate Pdf",
        "operationId": "generate_pdf_api_doctor_claims__claim_id__generate_pdf_post",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Claim Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratePdfResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doctor/claims/{claim_id}/mark-printed": {
      "post": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "Mark Printed",
        "operationId": "mark_printed_api_doctor_claims__claim_id__mark_printed_post",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Claim Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doctor/claims/{claim_id}/cancel": {
      "post": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_doctor_claims__claim_id__cancel_post",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Claim Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doctor/claims/{claim_id}/reuse-for-template": {
      "post": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "Reuse For Template",
        "operationId": "reuse_for_template_api_doctor_claims__claim_id__reuse_for_template_post",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Claim Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReuseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReuseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doctor/claims/{claim_id}": {
      "get": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "Get Claim",
        "operationId": "get_claim_api_doctor_claims__claim_id__get",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Claim Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClaimCreate": {
        "properties": {
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "template_id": {
            "type": "integer",
            "title": "Template Id"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "template_id"
        ],
        "title": "ClaimCreate"
      },
...
      "ClaimOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "submission_no": {
            "type": "string",
            "title": "Submission No"
          },
          "clinic_id": {
            "type": "integer",
            "title": "Clinic Id"
          },
          "doctor_id": {
            "type": "integer",
            "title": "Doctor Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "template_id": {
            "type": "integer",
            "title": "Template Id"
          },
          "template_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Version"
          },
          "patient_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Name"
          },
          "ai_raw_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Raw Result"
          },
          "final_field_values": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Field Values"
          },
          "ai_token_usage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Token Usage"
          },
          "ai_process_time_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Process Time Ms"
          },
          "generated_pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Pdf Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "submission_no",
          "clinic_id",
          "doctor_id",
          "company_id",
          "template_id",
          "template_version",
          "patient_name",
          "ai_raw_result",
          "final_field_values",
          "ai_token_usage",
          "ai_process_time_ms",
          "generated_pdf_url",
          "status",
          "created_at"
        ],
        "title": "ClaimOut"
      },
...
      "FieldsUpdate": {
        "properties": {
          "final_field_values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Final Field Values"
          }
        },
        "type": "object",
        "required": [
          "final_field_values"
        ],
        "title": "FieldsUpdate"
      },
      "GeneratePdfResponse": {
        "properties": {
          "pdf_url": {
            "type": "string",
            "title": "Pdf Url"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "pdf_url",
          "generated_at"
        ],
        "title": "GeneratePdfResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MedicalRecordSubmit": {
        "properties": {
          "medical_record_text": {
            "type": "string",
            "title": "Medical Record Text"
          },
          "patient_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Name"
          }
        },
        "type": "object",
        "required": [
          "medical_record_text"
        ],
        "title": "MedicalRecordSubmit"
      },
      "Page_ClaimListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClaimListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ClaimListItem]"
      },
...
      "ReuseRequest": {
        "properties": {
          "new_template_id": {
            "type": "integer",
            "title": "New Template Id"
          }
        },
        "type": "object",
        "required": [
          "new_template_id"
        ],
        "title": "ReuseRequest"
      },
      "ReuseResponse": {
        "properties": {
          "submission_id": {
            "type": "integer",
            "title": "Submission Id"
          },
          "prefilled_fields": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Prefilled Fields"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          }
        },
        "type": "object",
        "required": [
          "submission_id",
          "prefilled_fields",
          "missing_fields"
        ],
        "title": "ReuseResponse"
      },
...
      "TemplateBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "template_name": {
            "type": "string",
            "title": "Template Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "template_name",
          "version",
          "page_count"
        ],
        "title": "TemplateBrief"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}